handle sdk updates with retry logic, expose utility for generating and snapshotting the build script
