+++ build_runner/lib/src/build_script_generate/build_script_generate.dart
@@ -17,6 +17,7 @@ import '../package_graph/build_config_overrides.dart';
 import 'builder_ordering.dart';
 
 const scriptLocation = '$entryPointDir/build.dart';
+const scriptSnapshotLocation = '$scriptLocation.snapshot';
 
 Future<String> generateBuildScript() => logTimedAsync(
     Logger('Entrypoint'), 'Generating build script', _generateBuildScript);
